Improved the efficiency of the as.regperiod_range.character by using C++ code.

+++ src/period.h
@@ -1,6 +1,6 @@
 class Period {
 public:
-    // the first and last subperiod after Christ, starting at 0 in the year 0:
-    int p;
+    // the number of subperiod after Christ, starting at 0 in the year 0:
+    double p;
     int freq;
 };
